+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added consent controls to media embeds
 * Made PHP8.1 minimum version
 * Refactored package file structure